[MOD] 내 프로필 정보 조회, 회원 정보 조회 기능 수정 - 팔로워수, 게시글수 추가

+++ src/main/java/com/modugarden/domain/user/dto/response/UserInfoResponseDto.java
@@ -13,8 +13,8 @@ public class UserInfoResponseDto {
     private String nickname;
     private UserAuthority userAuthority;
     private String profileImage;
-    private int followerCount;
-    private int postCount;
+    private Long followerCount;
+    private Long postCount;
     private List<String> categories;
     private boolean isFollow;
 }
